bugfix: asymmetry in fluidization
In icPartUrqmd.cpp, the index of nearest hydro cell was calculated
by truncating the float instead of rounding it to the nearest integer.
This was causing a bias which led to slightly asymmetric spacetime
rapidity profile of initial energy density. It translated into a
slightly asymmetric rapidity profile dN/dy of the final state hadrons.
The bug persisted in all calculations from year 2015 onwards.