Multihydro freezeout: added calculation of dSigma and u
From the function getEnergyDensity I created a function
getEnergyMomentumTensor, which reads Q's of all three fluids and returns
the energy-momentum tensor. This function is then used also in function
findFreezeout to get energy-momentum tensor at the corners of the cube.

So T^munu is calculated at the corners of the cube, the interpolated to
the center of each segment and diagonalized to obtain e and v.

Issues:
- the code stops at tau=2.6, probably by filling up the memory
- dSigma_mu * u^mu is negative in more than 50% cases, which shouldn't
be