Add condition for switching resulting velocity sign
We calculate the resulting velocity as an eigenvector of energy-momentum
tensor of all three fluids combined. However, eigenvectors can be
multiplied by any number and they still remains eigenvectors.

When studying velocity fields of all three fluids and resulting fluid,
we found that the resulting fluid needs to be sometimes switched to
opposite direction. The condition that the scalar product of the
resulting velocity and average velocity of all three fluids needs to be
positive seems to resolve this issue.