Calculate EtotSurf at mutual hypersurface and write it to console
As a test of energy conversion from fluids to particles, we created a
variable including energy flow through the particlization surface called
EtotSurf. Similar variable is used in 1fluid version of the code - in
fld class.

We added calculation of EtotSurf at mutual hypersurface for individual
fluids. In addition, we also calculate positive and negative
contributions to the energy flow.