Put incoming nuclei further from each other at initial state
to avoid overlapping their energy densities at the beginning of the
hydro. Along with that I needed to add condition for the exit of the
hydro that it cannot happen earlier than few fm/c. Otherwise the hydro
would stop right at the beginning because there would be no freezeout
hypersurface elements.