Modifications to be compatible with SMASH hadronic EoS in the final state
- Update grid of SMASH EoS
- Modify determination of whether corona cell is in Tail. Previously, a
cell was considered in the Tail if the energy density in any of the
corners was above a certain value. This is now modified to consider the
energy density in the center of the cell, obtained from interpolation of T^{0mu} and not from e. This value is also used to create the freezeout surface.