Change time of end of simulation from 5 to tau0+5
The simulation stops when the number of elements is zero and also
some time has passed. Setting tau0>5 caused the simulation to
stop at the first timestep since the interaction of nuclei did
not start yet. This commit changes that simulation have to take
at least 5 fm before it can stop.