Remove code duplications in VTK source code
This commit restructures a part of the VTK source code to get rid of
a few lines of unnecessary or duplicated code. The function
`write_vtk_tensor` was not restructured (e.g., by just calling
`write_vtk_scalar(file, h, quantity+std::to_string(i)+std::to_string(j));`
inside the two for-loops and implementing an else-if-statement for the
components of `pi` in `write_vtk_scalar`) because this led to an
increase in runtime of 20-25% w.r.t. leaving the code duplicate.