Remove VTK_cartesian key and corresponding code
The VTK_cartesian output included a few bugs which are quite complex
to solve due to the transformation t = \tau \cosh(\eta) which leads to
a spread of values over multiple Cartesian time files. Since this
output was more of a pet project and is not used, this commit removes
it entirely.