range of transverse smearing reduced in IcPartSMASH
Checks by Gabriele Inghirami have shown that it is safe to reduce the
range of cells for energy-momentum deposition from the initial state
hadrons. The reduction affects the initial state very little in
event-by-event scenarios, however it reduces the time to compute the
initial state in scenarios with averaged initial state.